refactor(backend): create a class for socket authentication
